% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfourspec.R
\encoding{UTF-8}
\name{calc_periodogram}
\alias{calc_periodogram}
\title{calc_periodogram}
\usage{
calc_periodogram(x, P_avg)
}
\arguments{
\item{x}{a numeric vector or univariate time series.}

\item{P_avg}{\code{double}, Mean Square Value of x, aka x's mean power.}
}
\value{
the spectral density estimation.
}
\description{
This function estimates the spectral density as the Fourier Transform of the
auto covariance function.
}
\examples{

}
\keyword{internal}
