% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfourspec.R
\encoding{UTF-8}
\name{plot.sfourspec}
\alias{plot.sfourspec}
\title{plot_spectrum}
\usage{
\method{plot}{sfourspec}(x, fft_type = "exp", x_axis = "f", y_axis = "fourier_line_spectrum", ...)
}
\arguments{
\item{x}{TODO}

\item{fft_type}{TODO}

\item{x_axis}{TODO}

\item{y_axis}{TODO}

\item{...}{TODO}
}
\value{

}
\description{
plot_spectrum
}
