% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfourspec.R
\encoding{UTF-8}
\name{plot_spetrum}
\alias{plot_spetrum}
\title{plot_spetrum}
\usage{
plot_spetrum(
  spec_x,
  fft_type = "exp",
  x_axis = "f",
  y_axis = "fourier_line_spectrum",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
plot_spetrum
}
