% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfourspec.R
\encoding{UTF-8}
\name{rebuild_signal}
\alias{rebuild_signal}
\title{rebuild_signal}
\usage{
rebuild_signal(spec_x, threshold = 0.8)
}
\arguments{
\item{spec_x}{an object of class \code{sfourspec}.}

\item{threshold}{the % of power in the original signal to keep in the rebuilt
signal}
}
\value{
a numeric vector with the rebuilt signal.
}
\description{
rebuild_signal
}
