% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfourspec.R
\encoding{UTF-8}
\name{rebuild_signal_harmonics}
\alias{rebuild_signal_harmonics}
\title{rebuild_signal_harmonics}
\usage{
rebuild_signal_harmonics(spec_x, Ts)
}
\arguments{
\item{spec_x}{TODO}

\item{Ts}{TODO}
}
\value{

}
\description{
rebuild_signal_harmonics
}
