% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfourspec-package.R
\docType{package}
\name{sfourspec-package}
\alias{sfourspec-package}
\alias{_PACKAGE}
\alias{sfourspec}
\title{sfourspec}
\description{
This package provides with some simple tools to make density spectra from
time signals.
}
\references{
Cowpertwait, Paul S. P., and Andrew V. Metcalfe. 2009.
    Introductory Time Series with R. 1st ed.
    233 Spring Street, New York, NY 10013, USA: Springer Science+Business Media, LLC.
    \href{http://www.springer.com/series/6991}{Springer Series}.
}
