% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfourspec.R
\encoding{UTF-8}
\name{spectral_analysis}
\alias{spectral_analysis}
\title{spectral_analysis}
\usage{
spectral_analysis(x)
}
\arguments{
\item{x}{a numeric vector or univariate time series.}
}
\value{
List with:
    \itemize{
        \item P_avg \code{double}, Mean Square Value of x, aka x's mean power.
        \item P_dc \code{double}, Square mean value of x, aka the mean power of the
        x's continuous component (dc).
        \item P_ac \code{double}, Variance of x, aka the mean power of the x's
        alternate component (ac).
        \item x_ef \code{double},, The effective value of x.
        \item four_exp \code{tibble}, containing the results obtained using a
        Fourier series with exponential form (linear combination of complex
        exponentials).
        \item four_cos_sin \code{tibble}, containing the results obtained using a
        Fourier series with sine / cosine form (linear combination of sines and cosines)..
        \item four_cos \code{tibble}, containing the results obtained using a
        Fourier series with cosine form (linear combination of cosines).
        \item x, the original signal.
    }
}
\description{
Simple Fourier spectra
}
\details{
\deqn{
\langle x[n] \rangle_{N} = \frac{1}{N} \sum_{\langle N \rangle}x[n]
}
}
\examples{
air_pass <- tibble(t = 1:length(AirPassengers), pass = as.numeric(AirPassengers))

lm_air_pass <- lm(pass ~ t, air_pass)
summary(lm_air_pass)

detrended_air_pass <- AirPassengers - fitted(lm_air_pass)
N <- length(detrended_air_pass)
ns <- 0:(N-1)

spec_x <- spectral_analysis(detrended_air_pass)

plot(spec_x)

}
