% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfourspec.R
\encoding{UTF-8}
\name{calc_periodogram}
\alias{calc_periodogram}
\title{calc_periodogram}
\usage{
calc_periodogram(x, P_avg)
}
\arguments{
\item{P_avg}{}
}
\value{

}
\description{
calc_periodogram
}
\keyword{internal}
