% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfourspec.R
\encoding{UTF-8}
\name{rebuild_signal}
\alias{rebuild_signal}
\title{rebuild_signal}
\usage{
rebuild_signal(spec_x, threshold = 0.8)
}
\arguments{
\item{spec_x}{TODO}

\item{threshold}{TODO}
}
\value{

}
\description{
rebuild_signal
}
