% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfourspec.R
\encoding{UTF-8}
\name{spectral_analysis}
\alias{spectral_analysis}
\title{spectral_analysis}
\usage{
spectral_analysis(x)
}
\arguments{
\item{x}{}
}
\value{
List with:
    \itemize{
        \item P_avg \code{Tibble}, Term - estimate table for the model
                     coefficients.
        \item P_dc \code{matrix} containing the samples of the
                          coefficients.
        \item P_ac \code{matrix} with quantiles from the samples of the
                       model coeficcients
        \item x_ef \code{Tibble}, Term - estimate table for the model
                     coefficients.
        \item four_exp \code{matrix} containing the samples of the
                          coefficients.
        \item four_cos_sin \code{matrix} with quantiles from the samples of the
                       model coeficcients

        \item four_cos \code{matrix} with quantiles from the samples of the
                       model coeficcients
    }
}
\description{
spectral_analysis
}
\examples{
N_0 <- 10       # PERIODO - u. de t. (p. ej., segundos)
w_0 <- 2*pi/N_0 # FRECUENCIA ANGULAR

# (muestreo cada milisegundo)
sampling_freq <- 1e-3

ns <- seq(0, N_0/sampling_freq-1, by = 1)
N <- length(ns)
x <- sin(w_0*ns*sampling_freq)

spec_x <- spectral_analysis(x)

}
