% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfourspec.R
\encoding{UTF-8}
\name{spectral_analysis}
\alias{spectral_analysis}
\title{spectral_analysis}
\usage{
spectral_analysis(x)
}
\arguments{
\item{x}{a numeric vector or univariate time series.}
}
\value{
List with:
    \itemize{
        \item P_avg \code{double}, Mean Square Value of x, aka x's mean power.
        \item P_dc \code{double}, Square mean value of x, aka the mean power of the
        x's continuous component (dc).
        \item P_ac \code{double}, Variance of x, aka the mean power of the x's
        alternate component (ac).
        \item x_ef \code{double},, The effective value of x.
        \item four_exp \code{tibble}, containing the results obtained using a
        Fourier series with exponential form (linear combination of complex
        exponentials).
        \item four_cos_sin \code{tibble}, containing the results obtained using a
        Fourier series with sine / cosine form (linear combination of sines and cosines)..
        \item four_cos \code{tibble}, containing the results obtained using a
        Fourier series with cosine form (linear combination of cosines).
        \item x, the original signal.
    }
}
\description{
Simple Fourier spectra
}
\details{
\deqn{
\langle x[n] \rangle_{N} = \frac{1}{N} \sum_{\langle N \rangle}x[n]
}
}
\examples{
N_0 <- 10       # PERIODO - u. de t. (p. ej., segundos)
w_0 <- 2*pi/N_0 # FRECUENCIA ANGULAR

# (muestreo cada milisegundo)
sampling_freq <- 1e-3

ns <- seq(0, N_0/sampling_freq-1, by = 1)
N <- length(ns)
x <- sin(w_0 * ns * sampling_freq)

spec_x <- spectral_analysis(x)

plot(spec_x)

}
